ASKAPSDP-2661 - New paramter USE_CLI to control whether we use the online systems.
If not, we do various checks for already-created metadata files. If 
they exist, we carry on. If not, we stop everything.

Had to make a small change to the footprint.py interface as well.

+++ Code/Components/Pipelines/bash-pipelines/current/scripts/bin/findBeamCentres.sh
@@ -29,15 +29,17 @@
 #
 
 tmpfp="${tmp}/listOfFootprints"
-module load askapcli
-footprint list > "$tmpfp"
-err=$?
-module unload askapcli
-if [ $err -ne 0 ]; then
-    echo "ERROR - the 'footprint' command failed. "
-    echo "        Full command:   footprint list"
-    echo "Exiting pipeline."
-    exit $err
+if [ "${USE_CLI}" == "true" ] && [ "${IS_BETA}" != "true" ]; then
+    module load askapcli
+    footprint list > "$tmpfp"
+    err=$?
+    module unload askapcli
+    if [ $err -ne 0 ]; then
+        echo "ERROR - the 'footprint' command failed. "
+        echo "        Full command:   footprint list"
+        echo "Exiting pipeline."
+        exit $err
+    fi
 fi
 
 NEED_BEAM_CENTRES=false
@@ -52,23 +54,40 @@ if [ "$DO_SCIENCE_FIELD" == "true" ] && [ "$NEED_BEAM_CENTRES" == "true" ]; then
     # For the non-BETA case, we use schedblock from the askapcli
     # module - this polls the online scheduling-block database
     if [ "${IS_BETA}" != "true" ]; then
-        
-        # Run schedblock to get footprint information (if present)
+
         sbinfo="${metadata}/schedblock-info-${SB_SCIENCE}.txt"
-        if [ ! -e "${sbinfo}" ] || [ "$(wc -l "$sbinfo" | awk '{print $1}')" -gt 1 ]; then
+        if [ "${USE_CLI}" != "true" ]; then
+
             if [ -e "${sbinfo}" ]; then
-                rm -f "$sbinfo"
+                echo "Reusing schedblock info results $sbinfo"
+            else
+
+                echo "The schedblock service is not available, and you don't have a pre-computed SB info file."
+                echo "    (did not find $sbinfo)"
+                echo "Halting the pipeline here as we cannot find the beam locations."
+                echo "You may want to use the IS_BETA option, giving the beam footprint information directly."
+                exit 1
             fi
-            module load askapcli
-            schedblock info -v -p "${SB_SCIENCE}" > "$sbinfo"
-            err=$?
-            module unload askapcli
-            if [ $err -ne 0 ]; then
-                echo "ERROR - the 'schedblock' command failed."
-                echo "        Full command:   schedblock info -v -p ${SB_SCIENCE}"
-                echo "Exiting pipeline."
-                exit $err
+
+        else
+            
+            # Run schedblock to get footprint information (if present)
+            if [ ! -e "${sbinfo}" ] || [ "$(wc -l "$sbinfo" | awk '{print $1}')" -gt 1 ]; then
+                if [ -e "${sbinfo}" ]; then
+                    rm -f "$sbinfo"
+                fi
+                module load askapcli
+                schedblock info -v -p "${SB_SCIENCE}" > "$sbinfo"
+                err=$?
+                module unload askapcli
+                if [ $err -ne 0 ]; then
+                    echo "ERROR - the 'schedblock' command failed."
+                    echo "        Full command:   schedblock info -v -p ${SB_SCIENCE}"
+                    echo "Exiting pipeline."
+                    exit $err
+                fi
             fi
+            
         fi
         defaultFPname=$(grep "%d.footprint.name" "${sbinfo}" | awk '{print $3}')
         defaultFPpitch=$(grep "%d.footprint.pitch" "${sbinfo}" | awk '{print $3}')
@@ -167,36 +186,22 @@ if [ "$DO_SCIENCE_FIELD" == "true" ] && [ "$NEED_BEAM_CENTRES" == "true" ]; then
         #   ACES for BETA data or old ASKAP data that doesn't have
         #   footprint information in the database.
         
-        # Check to see whether the footprint name is used by ASKAPCLI/footprint
-        beamFromCLI=true
-        if [ "${IS_BETA}" == "true" ] || [ "$(grep "$FP_NAME" "${tmpfp}")" == "" ]; then
-            beamFromCLI=false
-            if [ "${IS_BETA}" == "true" ]; then
-                echo "Using the ACES footprint.py tool for BETA data"
-            else
-                echo "Footprint name $FP_NAME is not recognised by askapcli/footprint. Using the ACES footprint.py tool"
-            fi
-            if [ "$(which footprint.py 2> "${tmp}/whchftprnt")" == "" ]; then
-	        # If we are here, footprint.py is not in our path. Give an
-	        # error message and turn off linmos
-                
-                if [ "${DO_MOSAIC}" == "true" ]; then
-	            echo "ERROR - Cannot find 'footprint.py', so cannot determine beam arrangement. 
-      Setting DO_MOSAIC=false."  
-	            DO_MOSAIC=false
-                fi
-                
-                if [ "$IMAGE_AT_BEAM_CENTRES" == "true" ]; then
-                    echo "ERROR - Cannot find 'footprint.py', so cannot set beam centres. 
-      Not running - you may want to find footprint.py or               change your config file."
-                    SUBMIT_JOBS=false
-                fi
-            fi
-        fi
-
         # define the output file as $footprintOut
         setFootprintFile
         
+        # Check to see whether the footprint name is used by ASKAPCLI/footprint
+        beamFromCLI=true
+        
+        if [ "${IS_BETA}" == "true" ]; then
+            beamFromCLI=false
+            echo "Using the ACES footprint.py tool for BETA data"
+        fi
+
+        if [ "${USE_CLI}" == "true" ] && [ "$(grep "$FP_NAME" "${tmpfp}")" == "" ]; then
+            beamFromCLI=false
+            echo "Footprint name $FP_NAME is not recognised by askapcli/footprint. Using the ACES footprint.py tool"
+        fi
+
         # If the footprint output file exists, we don't re-run the
         # footprint determination.
         # The only exception to that is if it exists but is empty - a previous footprint
@@ -209,7 +214,16 @@ if [ "$DO_SCIENCE_FIELD" == "true" ] && [ "$NEED_BEAM_CENTRES" == "true" ]; then
             fi
             echo "Writing footprint for field $FIELD to file $footprintOut"
 
-            if [ "$beamFromCLI" == "true" ]; then 
+            if [ "$beamFromCLI" == "true" ]; then
+
+                if [ "${USE_CLI}" != "true" ]; then
+                    # If the service is not available, and we don't have beam locations, stop everything.
+                    echo "The footprint service is not available, and you don't have a pre-computed footprint file."
+                    echo "Halting the pipeline here as we cannot find the beam locations."
+                    echo "You may want to use the IS_BETA option, giving the beam footprint information directly."
+                    exit 1
+                fi
+
                 # This uses the CLI tool "footprint" to set the footprint
                 footprintArgs="-d $ra,$dec -p $FP_PITCH"
                 if [ "$FP_PA" != "" ]; then
@@ -227,36 +241,54 @@ if [ "$DO_SCIENCE_FIELD" == "true" ] && [ "$NEED_BEAM_CENTRES" == "true" ]; then
                 fi
             else
                 # This case uses the ACES tool "footprint.py"
-                
-                # First, need to check that the footprint provided is
-                # valid (ie. recognised by footprint.py)
-                module load aces
-                invalidTest=$(footprint.py -n "${BEAM_FOOTPRINT_NAME}" 2>&1 | grep invalid)
-                module unload aces
-                if [ "${FP_NAME}" == "" ] || [ "${invalidTest}" != "" ]; then
-                    # We don't have a valid footprint name!
-                    echo "ERROR - Your requested footprint ${BEAM_FOOTPRINT_NAME} is not valid."
-                    if [ "${IMAGE_AT_BEAM_CENTRES}" == "true" ]; then
-                        echo "      Not running - change your config file."
-                        SUBMIT_JOBS=false
-                    else
-                        if [ "${DO_MOSAIC}" == "true" ]; then
-                            echo "      Setting DO_MOSAIC to false."
-                        fi
+                if [ "$(which footprint.py 2> "${tmp}/whchftprnt")" == "" ]; then
+	            # If we are here, footprint.py is not in our path. Give an
+	            # error message and turn off linmos
+                    
+                    if [ "${DO_MOSAIC}" == "true" ]; then
+	                echo "ERROR - Cannot find 'footprint.py', so cannot determine beam arrangement. 
+      Setting DO_MOSAIC=false."  
+	                DO_MOSAIC=false
                     fi
                     
+                    if [ "$IMAGE_AT_BEAM_CENTRES" == "true" ]; then
+                        echo "ERROR - Cannot find 'footprint.py', so cannot set beam centres. 
+      Not running - you may want to find footprint.py or               change your config file."
+                        SUBMIT_JOBS=false
+                    fi
+
                 else
-                    # Use the function defined in utils.sh to set the arguments to footprint.py
-                    setFootprintArgs
+                
+                    # Next, need to check that the footprint provided is
+                    # valid (ie. recognised by footprint.py)
                     module load aces
-                    footprint.py "$footprintArgs" -r "$ra,$dec" > "${footprintOut}"
-                    err=$?
+                    invalidTest=$(footprint.py -n "${BEAM_FOOTPRINT_NAME}" 2>&1 | grep invalid)
                     module unload aces
-                    if [ $err -ne 0 ]; then
-                        echo "ERROR - the 'footprint.py' command failed. "
-                        echo "        Full command:   footprint.py $footprintArgs -r \"$ra,$dec\""
-                        echo "Exiting pipeline."
-                        exit $err
+                    if [ "${FP_NAME}" == "" ] || [ "${invalidTest}" != "" ]; then
+                        # We don't have a valid footprint name!
+                        echo "ERROR - Your requested footprint ${BEAM_FOOTPRINT_NAME} is not valid."
+                        if [ "${IMAGE_AT_BEAM_CENTRES}" == "true" ]; then
+                            echo "      Not running - change your config file."
+                            SUBMIT_JOBS=false
+                        else
+                            if [ "${DO_MOSAIC}" == "true" ]; then
+                                echo "      Setting DO_MOSAIC to false."
+                            fi
+                        fi
+                        
+                    else
+                        # Use the function defined in utils.sh to set the arguments to footprint.py
+                        setFootprintArgs
+                        module load aces
+                        footprint.py $footprintArgs -r "$ra,$dec" > "${footprintOut}"
+                        err=$?
+                        module unload aces
+                        if [ $err -ne 0 ]; then
+                            echo "ERROR - the 'footprint.py' command failed. "
+                            echo "        Full command:   footprint.py $footprintArgs -r \"$ra,$dec\""
+                            echo "Exiting pipeline."
+                            exit $err
+                        fi
                     fi
                 fi
             fi
